+++ database/migrations/2024_01_01_000023_create_sales_table.php
@@ -10,10 +10,10 @@ return new class extends Migration
     {
         Schema::create('sales', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('stock_card_id');
+            $table->unsignedBigInteger('stock_card_id');
             $table->enum('type', ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10'])->comment('1 = Telefon↲2 = Aksesuar↲3 = Teknik Ürün↲4 = Diğer');
-            $table->integer('stock_card_movement_id');
-            $table->integer('invoice_id');
+            $table->unsignedBigInteger('stock_card_movement_id');
+            $table->unsignedBigInteger('invoice_id');
             $table->decimal('sale_price', 10, 2);
             $table->decimal('customer_price', 10, 2);
             $table->decimal('cash_price', 10, 2);
@@ -27,9 +27,9 @@ return new class extends Migration
             $table->string('serial');
             $table->decimal('discount', 10, 2);
             $table->datetime('creates_dates');
-            $table->integer('technical_service_person_id')->nullable();
+            $table->unsignedBigInteger('technical_service_person_id')->nullable();
             $table->decimal('base_cost_price', 10, 2);
-            $table->integer('delivery_personnel')->nullable();
+            $table->unsignedBigInteger('delivery_personnel')->nullable();
             $table->timestamps();
             
             $table->foreign('seller_id')->references('id')->on('sellers');
